Add test for diffing files above normal 10MiB limit
Added two new test files of >10MiB in size that contain garbage
printable ASCII characters. These are Test1_Huge and Test2_Huge.

Added new "patch huge" command to Test.bat that performs a binary
diffing of the new "huge" files, then applies the patch and tests the
result using FC.

The user is asked to confirm the operation because it takes forever.